wip: Refactor DiscoveryCache to tolerate inter-service errors
