+++ registrar/apps/api/v1/tests/test_views.py
@@ -25,7 +25,10 @@ from registrar.apps.api.constants import ENROLLMENT_WRITE_MAX_SIZE
 from registrar.apps.api.tests.mixins import AuthRequestMixin, TrackTestMixin
 from registrar.apps.core import permissions as perms
 from registrar.apps.core.constants import PROGRAM_CACHE_KEY_TPL
-from registrar.apps.core.data import DiscoveryCourseRun, DiscoveryProgram
+from registrar.apps.core.discovery_cache import (
+    DiscoveryCourseRun,
+    DiscoveryProgram,
+)
 from registrar.apps.core.filestore import (
     get_enrollment_uploads_filestore,
     get_program_reports_filestore,
@@ -1130,7 +1133,7 @@ class ProgramEnrollmentWriteMixin(object):
         mock_response = mock.Mock()
         mock_response.status_code = 404
         error = requests.exceptions.HTTPError(response=mock_response)
-        with mock.patch('registrar.apps.core.data._make_request', side_effect=error):
+        with mock.patch('registrar.apps.core.discovery_cache.make_request', side_effect=error):
             response = self.request(
                 self.method,
                 'programs/masters-in-cs/enrollments/',
