+++ registrar/apps/core/tasks.py
@@ -31,7 +31,7 @@ def debug_user_task(self, user_id, text):
     UserTaskArtifact.objects.create(status=self.status, text=text)
 
 
-def _get_program(job_id, program_key):
+def get_program(job_id, program_key):
     """
     Load a Program by key. Fails job and returns None if key invalid.
     """
